requirements.txt
Django==3.2.6
gunicorn==20.1.0
dj-database-url==0.5.0
djangorestframework==3.13.1
psycopg2-binary==2.9.2
whitenoise==5.3.0
coreapi==2.3.3
drf-yasg==1.21.4
djangorestframework-simplejwt==5.2.1
django-cors-headers==3.13.0

apscheduler==3.9.1
django-dbbackup==4.0.2
mega.py==1.0.8